refactor(adoc-core): unit 10 — validate.rs → infrastructure/validate/mod.rs
Move validation pass under infrastructure. File already used
crate::domain::* paths. Update compile.rs import from crate::validate
to crate::infrastructure::validate.

Refs #27.
