+++ crates/adoc-core/src/infrastructure/mod.rs
@@ -3,3 +3,4 @@ pub(crate) mod artifact;
 pub(crate) mod parser;
 pub(crate) mod render;
 pub(crate) mod source;
+pub(crate) mod validate;
