+++ crates/adoc-core/src/infrastructure/validate/mod.rs
